+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # CHANGELOG   
 
 
-## Upcoming   
-MINOR: style show more btn responsive #145
+## Upcoming 1.3.0   
+- **NEW**: MQTT Config Tab on device config page #101
+  - ENHANCEMENT: Only changed config values will be send to the device
+- MINOR: style show more btn responsive #145
+
 
 ## v1.2.1
 **HOTFIX**: Fix edit device #147